Modify flowchart to reflect addition of erogaki-classify and -detect modules
The hent-AI-erogaki-wrapper will be replaced by erogaki-detect, which will
handle detection and masking of censored regions.
DCP will also handle mosaic decensoring.
The flowchart has been simplified to not distinguish between bar and mosaic
decensoring when not relevant for the execution flow between modules.
As mosaic decensoring has to be handled differently than bar decensoring
in terms of the data flow the redis_flows.md will be updated accordingly.
Additionally a new module "erogaki-classify" is planned to handle automatic
detection if an image is censored and which type of censor is applied.